+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## new
+* change: The parameters of ToBeanOrNull and ToJSONObjectOrNull are now marked as NotNull
+
+
 ## v1.0.1
 * downgrade: source and target Compatibility downgrade to 1.7, kotlinOptions.jvmTarget downgrade to 1.6